+++ openssl/crypto/asn1/i2d_s_pr.c
@@ -97,7 +97,7 @@ int i2d_DSAPrivateKey(DSA *a, unsigned char **pp)
 		}
 
 	t=ASN1_object_size(1,tot,V_ASN1_SEQUENCE);
-	if (pp == NULL) return(t);
+	if (pp == NULL) return((int)t);
 
 	p= *pp;
 	ASN1_put_object(&p,1,tot,V_ASN1_SEQUENCE,V_ASN1_UNIVERSAL);
@@ -118,6 +118,6 @@ int i2d_DSAPrivateKey(DSA *a, unsigned char **pp)
 		}
 	OPENSSL_free(bs.data);
 	*pp=p;
-	return(t);
+	return((int)t);
 	}
 #endif
