+++ openssl/crypto/bio/bss_null.c
@@ -145,6 +145,6 @@ static int null_gets(BIO *bp, char *buf, int size)
 static int null_puts(BIO *bp, const char *str)
 	{
 	if (str == NULL) return(0);
-	return(strlen(str));
+	return((int)strlen(str));
 	}
