+++ openssl/crypto/rand/randfile.c
@@ -100,7 +100,7 @@ int RAND_load_file(const char *file, long bytes)
 
 	i=stat(file,&sb);
 	/* If the state fails, put some crap in anyway */
-	RAND_add(&sb,sizeof(sb),0);
+	RAND_add(&sb,sizeof(sb),(double)0);
 	if (i < 0) return(0);
 	if (bytes == 0) return(ret);
 
@@ -115,7 +115,7 @@ int RAND_load_file(const char *file, long bytes)
 		i=fread(buf,1,n,in);
 		if (i <= 0) break;
 		/* even if n != i, use the full array */
-		RAND_add(buf,n,i);
+		RAND_add(buf,n,(double)i);
 		ret+=i;
 		if (bytes > 0)
 			{
