+++ openssl/crypto/x509v3/v3_purp.c
@@ -123,7 +123,7 @@ int X509_check_purpose(X509 *x, int id, int ca)
 int X509_PURPOSE_get_count(void)
 {
 	if(!xptable) return X509_PURPOSE_COUNT;
-	return sk_X509_PURPOSE_num(xptable) + X509_PURPOSE_COUNT;
+	return (int)(sk_X509_PURPOSE_num(xptable) + X509_PURPOSE_COUNT);
 }
 
 X509_PURPOSE * X509_PURPOSE_get0(int idx)
@@ -155,7 +155,7 @@ int X509_PURPOSE_get_by_id(int purpose)
 	if(!xptable) return -1;
 	idx = sk_X509_PURPOSE_find(xptable, &tmp);
 	if(idx == -1) return -1;
-	return idx + X509_PURPOSE_COUNT;
+	return (int)(idx + X509_PURPOSE_COUNT);
 }
 
 int X509_PURPOSE_add(int id, int trust, int flags,
