+++ openssl/ssl/s2_srvr.c
@@ -163,7 +163,7 @@ int ssl2_accept(SSL *s)
 	void (*cb)()=NULL;
 	int new_state,state;
 
-	RAND_add(&l,sizeof(l),0);
+	RAND_add(&l,sizeof(l),(double)0);
 	ERR_clear_error();
 	clear_sys_error();
