+++ openssl/ssl/s3_pkt.c
@@ -539,7 +539,7 @@ int ssl3_write_bytes(SSL *s, int type, const void *buf_, int len)
 			 * in ciphersuites with known-IV weakness: */
 			s->s3->empty_fragment_done = 0;
 			
-			return tot+i;
+			return (int)(tot+i);
 			}
 
 		n-=i;
@@ -681,7 +681,7 @@ static int do_ssl3_write(SSL *s, int type, const unsigned char *buf,
 		/* we are in a recursive call;
 		 * just return the length, don't write out anything here
 		 */
-		return wr->length;
+		return (int)(wr->length);
 		}
 
 	/* now let's set up wb */
@@ -807,7 +807,7 @@ int ssl3_read_bytes(SSL *s, int type, unsigned char *buf, int len, int peek)
 		/* move any remaining fragment bytes: */
 		for (k = 0; k < s->s3->handshake_fragment_len; k++)
 			s->s3->handshake_fragment[k] = *src++;
-		return n;
+		return (int)n;
 	}
 
 	/* Now s->s3->handshake_fragment_len == 0 if type == SSL3_RT_HANDSHAKE. */
@@ -890,7 +890,7 @@ start:
 				rr->off=0;
 				}
 			}
-		return(n);
+		return((int)n);
 		}
