+++ openssl/ssl/s3_srvr.c
@@ -169,7 +169,7 @@ int ssl3_accept(SSL *s)
 	int ret= -1;
 	int new_state,state,skip=0;
 
-	RAND_add(&Time,sizeof(Time),0);
+	RAND_add(&Time,sizeof(Time),(double)0);
 	ERR_clear_error();
 	clear_sys_error();
